Replace integers with fixed-with types
Fixed-width types reduced ambiguity as to the intended bahviour of the
code.
size_t is used for the length of arrays as well.

+++ main.h
@@ -8,14 +8,16 @@
 #define MAIN_H
 
 #include "fractions.h"
+#include "stddef.h"
 
-int count_char_in_string(const char, const char *);
-void pp_matrix(fraction **const, const int, const int);
-int find_greatest_value_in_column(fraction **const, const int, const int);
-void substract_lines_in_place(fraction *const, fraction *const, const int);
-void multiply_line_in_place(fraction *const, const fraction, const int);
-void triangularise(fraction **const, const int, const int);
-void print_results(fraction **const, const int, const int);
-void gaussian_elimination(fraction **const, const int, const int);
+size_t count_char_in_string(const char, const char *);
+void pp_matrix(fraction **const, const size_t, const size_t);
+size_t find_greatest_value_in_column(fraction **const, const size_t,
+                                     const size_t);
+void substract_lines_in_place(fraction *const, fraction *const, const size_t);
+void multiply_line_in_place(fraction *const, const fraction, const size_t);
+void triangularise(fraction **const, const size_t, const size_t);
+void print_results(fraction **const, const size_t, const size_t);
+void gaussian_elimination(fraction **const, const size_t, const size_t);
 
 #endif /* MAIN_H */
